wip : fix things

+++ MovieReviewSite.Core/Repositories/Genre/ModifyGenreForMovie.cs
@@ -23,6 +23,10 @@ public partial class GenreRepository
 
     public async Task AddGenreByMovieId(MovieGenreRequest dto)
     {
+        if (dto.ModifierRoleCode != 1 || dto.ModifierRoleCode != 3)
+        {
+            throw new ArgumentException("user is not authorized to make these changes!");
+        }
         var movieGenreList = await GetGenreByMovieId(dto.MovieId);
         var movie = await _context.Movies.Where(m => m.Id == dto.MovieId).SingleOrDefaultAsync();
         if (movieGenreList.Any(m => dto.GenreId == m.Id))
@@ -41,6 +45,10 @@ public partial class GenreRepository
 
     public async Task RemoveGenreByMovieId(MovieGenreRequest dto)
     {
+        if (dto.ModifierRoleCode != 1 || dto.ModifierRoleCode != 3)
+        {
+            throw new ArgumentException("user is not authorized to make these changes!");
+        }
         var movieGenre = await _context.MovieGenres.Where(mg => mg.MovieId == dto.MovieId && mg.GenreId == dto.GenreId)
             .SingleOrDefaultAsync();
         _context.MovieGenres.Remove(movieGenre!);
